move common resource fixtures to resource level conftest
