+++ h2integrate/resource/wind/test/conftest.py
@@ -1,1 +1,3 @@
+from h2integrate.resource.test.conftest import site_config, plant_simulation  # noqa: F401
+
 from test.conftest import temp_dir, pytest_collection_modifyitems  # noqa: F401
